backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-CORS==4.0.0
Flask-Migrate==4.0.5
Werkzeug==3.0.1
python-dotenv==1.0.0
email-validator==2.1.0
python-dateutil==2.8.2
gunicorn==21.2.0
psycopg2-binary==2.9.9

